split off jpgC into JPEG specific files
