+++ libheif/box.cc
@@ -23,6 +23,7 @@
 #include "box.h"
 #include "security_limits.h"
 #include "nclx.h"
+#include "jpeg.h"
 #include "jpeg2000.h"
 #include "hevc.h"
 #include "mask_image.h"
@@ -3295,38 +3296,6 @@ std::string Box_url::dump(Indent& indent) const
   sstr << indent << "location: " << m_location << "\n";
 
   return sstr.str();
-}
-
-
-std::string Box_jpgC::dump(Indent& indent) const
-{
-  std::ostringstream sstr;
-  sstr << Box::dump(indent);
-
-  sstr << indent << "num bytes: " << m_data.size() << "\n";
-
-  return sstr.str();
-}
-
-
-Error Box_jpgC::write(StreamWriter& writer) const
-{
-  size_t box_start = reserve_box_header_space(writer);
-
-  writer.write(m_data);
-
-  prepend_header(writer, box_start);
-
-  return Error::Ok;
-}
-
-
-Error Box_jpgC::parse(BitstreamRange& range)
-{
-  size_t nBytes = range.get_remaining_bytes();
-  m_data.resize(nBytes);
-  range.read(m_data.data(), nBytes);
-  return range.get_error();
 }
