+++ libheif/file.cc
@@ -22,6 +22,7 @@
 #include "libheif/box.h"
 #include "libheif/heif.h"
 #include "libheif/jpeg2000.h"
+#include "libheif/jpeg.h"
 
 #include <cstdint>
 #include <fstream>
